pyproject.toml
[project]
name = "winxtract"
version = "0.1.0"
description = "Independent multi-source public B2C scraping engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "playwright>=1.53.0",
  "typer>=0.16.0",
  "pyyaml>=6.0.2",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.11.0",
  "sqlalchemy>=2.0.41",
  "psycopg[binary]>=3.2.0",
  "httpx>=0.28.1",
  "tenacity>=9.1.2",
  "structlog>=25.4.0",
  "phonenumbers>=9.0.8",
  "email-validator>=2.2.0",
  "openpyxl>=3.1.5",
]

[project.optional-dependencies]
api = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "jinja2>=3.1.6",
  "python-multipart>=0.0.20",
]
dev = [
  "pytest>=8.4.1",
  "pytest-asyncio>=1.1.0",
  "ruff>=0.12.9",
]

[project.scripts]
winxtract = "winxtract.cli:app"

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["winxtract*"]

[tool.setuptools.package-data]
winxtract = ["ui/templates/*.html", "ui/static/*.css"]
